centralize config file logic in runtime
