+++ e2e/sveltekit/houdini.config.js
@@ -1,3 +1,5 @@
+/// <references types="houdini-svelte">
+
 /** @type {import('houdini').ConfigFile} */
 const config = {
   client: './src/lib/graphql/houdiniClient',
