+++ packages/houdini-svelte/src/runtime/stores/pagination/offset.ts
@@ -1,7 +1,7 @@
 import { getCache } from '$houdini/runtime'
+import { ConfigFile } from '$houdini/runtime/lib/config'
 import { deepEquals } from '$houdini/runtime/lib/deepEquals'
 import { executeQuery } from '$houdini/runtime/lib/network'
-import { ConfigFile } from '$houdini/runtime/lib/types'
 import { GraphQLObject, QueryArtifact, QueryResult } from '$houdini/runtime/lib/types'
 
 import { getSession } from '../../session'
