+++ packages/houdini/src/test/index.ts
@@ -4,7 +4,8 @@ import path from 'path'
 
 import { runPipeline } from '../codegen'
 import { Config, fs, CollectedGraphQLDocument } from '../lib'
-import { ConfigFile, ArtifactKind } from '../runtime/lib/types'
+import { ConfigFile } from '../runtime/lib/config'
+import { ArtifactKind } from '../runtime/lib/types'
 
 export function testConfigFile(config: Partial<ConfigFile> = {}): ConfigFile {
 	return {
